CHanges for wallet transfer logic
